@@ -1,78 +1,87 @@
-import React, { Component } from 'react'
-import { login } from './UserFunctions'
+import React, { useState } from 'react';
+import { useHistory } from 'react-router-dom';
+import { login } from './UserFunctions';
 
-class Login extends Component {
-  constructor() {
-    super()
-    this.state = {
-      email: '',
-      password: '',
-      errors: {}
-    }
+function Login() {
+	const [email, setEmail] = useState('');
+	const [password, setPassword] = useState('');
+	const [error, setError] = useState('');
+	const [errorMessage, setErrorMessage] = useState(false);
 
-    this.onChange = this.onChange.bind(this)
-    this.onSubmit = this.onSubmit.bind(this)
-  }
+	const history = useHistory();
 
-  onChange(e) {
-    this.setState({ [e.target.name]: e.target.value })
-  }
-  onSubmit(e) {
-    e.preventDefault()
+	function validateForm() {
+		if (email && password) {
+		} else {
+			setError(true);
+			setErrorMessage('Ambos datos son obligatorios');
+		}
+	}
 
-    const user = {
-      email: this.state.email,
-      password: this.state.password
-    }
+	const sendData = (e) => {
+		e.preventDefault();
 
-    login(user).then(res => {
-      if (res) {
-        this.props.history.push(`/profile`)
-      }
-    })
-  }
+		let user = {
+			email,
+			password,
+		};
 
-  render() {
-    return (
-      <div className="container">
-        <div className="row">
-          <div className="col-md-6 mt-5 mx-auto">
-            <form noValidate onSubmit={this.onSubmit}>
-              <h1 className="h3 mb-3 font-weight-normal text-center">Inicia sesión</h1>
-              <div className="form-group">
-                <label htmlFor="email">Correo</label>
-                <input
-                  type="email"
-                  className="form-control"
-                  name="email"
-                  placeholder="Ingresa tu correo"
-                  value={this.state.email}
-                  onChange={this.onChange}
-                />
-              </div>
-              <div className="form-group">
-                <label htmlFor="password">Contraseña</label>
-                <input
-                  type="password"
-                  className="form-control"
-                  name="password"
-                  placeholder="Ingresa tu contraseña"
-                  value={this.state.password}
-                  onChange={this.onChange}
-                />
-              </div>
-              <button
-                type="submit"
-                className="btn btn-lg btn-primary btn-block"
-              >
-                Ingresar
-              </button>
-            </form>
-          </div>
-        </div>
-      </div>
-    )
-  }
+		login(user).then((res) => {
+			validateForm();
+			if (res) {
+				history.push('/');
+			}
+			console.log(`La respuesta de Login es ${res}`);
+		});
+	};
+
+	return (
+		<div className='d-flex align-content-center h-100'>
+			<div className='m-auto'>
+				<form onSubmit={sendData}>
+					<h1 className='h3 mb-3 font-weight-normal text-center'>
+						Inicia sesión
+					</h1>
+					{error && (
+						<label className='elementToFadeInAndOut alert alert-danger w-100 text-center'>
+							{errorMessage}
+						</label>
+					)}
+					<div className='form-group'>
+						<label htmlFor='email'>Correo</label>
+						<input
+							type='email'
+							className='form-control'
+							name='email'
+							placeholder='Ingresa tu correo'
+							value={email}
+							onChange={(e) => setEmail(e.target.value)}
+							// required
+						/>
+					</div>
+					<div className='form-group'>
+						<label htmlFor='password'>Contraseña</label>
+						<input
+							type='password'
+							className='form-control'
+							name='password'
+							placeholder='Ingresa tu contraseña'
+							value={password}
+							onChange={(e) => setPassword(e.target.value)}
+							// required
+						/>
+					</div>
+					<button
+						type='submit'
+						// disabled={!validateForm()}
+						className='btn btn-lg btn-primary btn-block'
+					>
+						Ingresar
+					</button>
+				</form>
+			</div>
+		</div>
+	);
 }
 
-export default Login
+export default Login;
