@@ -65,7 +65,7 @@ users.post('/login', (req, res) => {
 			}
 		})
 		.catch((err) => {
-			res.status(400).json({ error: err });
+      res.status(400).json({ error: err });
 		});
 });
 
